requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.10.4
anyio==3.6.2
apache-airflow==2.5.3
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
apispec==3.3.2
apprise==1.3.0
argcomplete==3.0.8
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgi-lifespan==2.1.0
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
blinker==1.4
cachelib==0.9.0
cachetools==5.3.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cli-helpers==2.3.0
click==8.1.3
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.4
colorlog==4.8.0
comm==0.1.2
command-not-found==0.3
configobj==5.0.8
ConfigUpdater==3.1.1
connexion==2.14.2
coolname==2.2.0
cron-descriptor==1.2.35
croniter==1.3.14
cryptography==40.0.2
dask==2023.4.1
dateparser==1.1.8
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distributed==2023.4.1
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
docker==6.0.1
docutils==0.19
email-validator==1.3.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.1
fastjsonschema==2.16.2
Flask==2.2.4
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud==0.34.0
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
graphviz==0.20.1
greenlet==2.0.2
griffe==0.27.1
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httplib2==0.20.2
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.1
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyring==23.5.0
keyrings.alt==4.2.0
kubernetes==26.1.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
locket==1.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==2.0.4
more-itertools==8.10.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.0
orjson==3.8.11
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pgcli==3.5.0
pgspecial==2.0.1
pickleshare==0.7.5
platformdirs==2.6.2
pluggy==1.0.0
prefect==1.4.1
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.22.3
psutil==5.9.4
psycopg==3.1.8
psycopg2==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.19.3
pyspark==3.4.0
python-apt==2.3.0+ubuntu2.1
python-box==7.0.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
readchar==4.0.5
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rsa==4.9
SecretStorage==3.3.1
Send2Trash==1.8.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.26.1
systemd-python==234
tabulate==0.9.0
tblib==1.7.0
tenacity==8.2.2
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.0
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.8.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==27.11.2
uc-micro-py==1.0.1
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.22.0
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
websockets==11.0.2
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zict==3.0.0
zipp==1.0.0
